qrg: Refactor qrcode package
