+++ pkg/qrcodegenerator/grpc_end_device_generator.go
@@ -19,7 +19,6 @@ import (
 
 	pbtypes "github.com/gogo/protobuf/types"
 	qrcodegen "github.com/skip2/go-qrcode"
-	"go.thethings.network/lorawan-stack/v3/pkg/qrcode"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 )
 
@@ -28,7 +27,7 @@ type endDeviceQRCodeGeneratorServer struct {
 }
 
 func (s *endDeviceQRCodeGeneratorServer) GetFormat(ctx context.Context, req *ttnpb.GetQRCodeFormatRequest) (*ttnpb.QRCodeFormat, error) {
-	format := qrcode.GetEndDeviceFormat(req.FormatId)
+	format := s.QRG.qrCode.GetEndDeviceFormat(req.FormatId)
 	if format == nil {
 		return nil, errFormatNotFound.New()
 	}
@@ -39,14 +38,14 @@ func (s *endDeviceQRCodeGeneratorServer) ListFormats(ctx context.Context, _ *pbt
 	res := &ttnpb.QRCodeFormats{
 		Formats: make(map[string]*ttnpb.QRCodeFormat),
 	}
-	for k, f := range qrcode.GetEndDeviceFormats() {
+	for k, f := range s.QRG.qrCode.GetEndDeviceFormats() {
 		res.Formats[k] = f.Format()
 	}
 	return res, nil
 }
 
 func (s *endDeviceQRCodeGeneratorServer) Generate(ctx context.Context, req *ttnpb.GenerateEndDeviceQRCodeRequest) (*ttnpb.GenerateQRCodeResponse, error) {
-	formatter := qrcode.GetEndDeviceFormat(req.FormatId)
+	formatter := s.QRG.qrCode.GetEndDeviceFormat(req.FormatId)
 	if formatter == nil {
 		return nil, errFormatNotFound.New()
 	}
