+++ pkg/qrcodegenerator/grpc_end_device_generator_test.go
@@ -25,8 +25,8 @@ import (
 	"go.thethings.network/lorawan-stack/v3/pkg/component"
 	componenttest "go.thethings.network/lorawan-stack/v3/pkg/component/test"
 	"go.thethings.network/lorawan-stack/v3/pkg/log"
-	"go.thethings.network/lorawan-stack/v3/pkg/qrcode"
 	. "go.thethings.network/lorawan-stack/v3/pkg/qrcodegenerator"
+	"go.thethings.network/lorawan-stack/v3/pkg/qrcodegenerator/qrcode"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 	"go.thethings.network/lorawan-stack/v3/pkg/types"
 	"go.thethings.network/lorawan-stack/v3/pkg/util/test"
@@ -37,7 +37,9 @@ func TestGenerateEndDeviceQRCode(t *testing.T) {
 	a := assertions.New(t)
 	ctx := log.NewContext(test.Context(), test.GetLogger(t))
 
-	qrcode.RegisterEndDeviceFormat("test", new(mockFormat))
+	qrCode := qrcode.New(ctx)
+
+	qrCode.RegisterEndDeviceFormat("test", new(mockFormat))
 
 	c := componenttest.NewComponent(t, &component.Config{})
 	test.Must(New(c, &Config{}))
