+++ pkg/qrcodegenerator/qrcode/enddevice/lora_alliance_tr005d3_test.go
@@ -12,26 +12,26 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package qrcode_test
+package enddevice_test
 
 import (
 	"testing"
 
 	"github.com/smartystreets/assertions"
 	"go.thethings.network/lorawan-stack/v3/pkg/errors"
-	. "go.thethings.network/lorawan-stack/v3/pkg/qrcode"
+	. "go.thethings.network/lorawan-stack/v3/pkg/qrcodegenerator/qrcode/enddevice"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 	"go.thethings.network/lorawan-stack/v3/pkg/types"
 	"go.thethings.network/lorawan-stack/v3/pkg/util/test"
 	"go.thethings.network/lorawan-stack/v3/pkg/util/test/assertions/should"
 )
 
-func TestLoRaAllianceTR005(t *testing.T) {
+func TestLoRaAllianceTR005Draft3(t *testing.T) {
 	t.Run("Encode", func(t *testing.T) {
 		for _, tc := range []struct {
 			Name     string
 			Device   ttnpb.EndDevice
-			Expected LoRaAllianceTR005
+			Expected LoRaAllianceTR005Draft3
 		}{
 			{
 				Name: "Simple",
@@ -44,16 +44,16 @@ func TestLoRaAllianceTR005(t *testing.T) {
 						Value: "ABCD",
 					},
 				},
-				Expected: LoRaAllianceTR005{
-					JoinEUI:    types.EUI64{0x70, 0xb3, 0xd5, 0x7e, 0xd0, 0x00, 0x00, 0x00},
-					DevEUI:     types.EUI64{0x1, 0x2, 0x3, 0x4, 0x5, 0x6, 0x7, 0x8},
-					OwnerToken: "ABCD",
+				Expected: LoRaAllianceTR005Draft3{
+					JoinEUI:              types.EUI64{0x70, 0xb3, 0xd5, 0x7e, 0xd0, 0x00, 0x00, 0x00},
+					DevEUI:               types.EUI64{0x1, 0x2, 0x3, 0x4, 0x5, 0x6, 0x7, 0x8},
+					DeviceValidationCode: "ABCD",
 				},
 			},
 		} {
 			t.Run(tc.Name, func(t *testing.T) {
 				a := assertions.New(t)
-				var res LoRaAllianceTR005
+				var res LoRaAllianceTR005Draft3
 				err := res.Encode(&tc.Device)
 				if !a.So(err, should.BeNil) {
 					t.FailNow()
@@ -68,13 +68,13 @@ func TestLoRaAllianceTR005(t *testing.T) {
 			Name           string
 			Data           []byte
 			CanonicalData  []byte
-			Expected       LoRaAllianceTR005
+			Expected       LoRaAllianceTR005Draft3
 			ErrorAssertion func(t *testing.T, err error) bool
 		}{
 			{
 				Name: "Simple",
-				Data: []byte("LW:D0:42FFFFFFFFFFFFFF:4242FFFFFFFFFFFF:42FFFF42"),
-				Expected: LoRaAllianceTR005{
+				Data: []byte("URN:DEV:LW:42FFFFFFFFFFFFFF_4242FFFFFFFFFFFF_42FFFF42"),
+				Expected: LoRaAllianceTR005Draft3{
 					JoinEUI:  types.EUI64{0x42, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff},
 					DevEUI:   types.EUI64{0x42, 0x42, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff},
 					VendorID: [2]byte{0x42, 0xff},
@@ -83,28 +83,27 @@ func TestLoRaAllianceTR005(t *testing.T) {
 			},
 			{
 				Name: "Extensions",
-				Data: []byte("LW:D0:42FFFFFFFFFFFFFF:4242FFFFFFFFFFFF:42FFFF42:CCHECKSUM:O0102:SSERIAL:PPROPRIETARY"),
-				Expected: LoRaAllianceTR005{
-					JoinEUI:      types.EUI64{0x42, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff},
-					DevEUI:       types.EUI64{0x42, 0x42, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff},
-					VendorID:     [2]byte{0x42, 0xff},
-					ModelID:      [2]byte{0xff, 0x42},
-					Checksum:     "CHECKSUM",
-					OwnerToken:   "0102",
-					SerialNumber: "SERIAL",
-					Proprietary:  "PROPRIETARY",
+				Data: []byte("URN:DEV:LW:42FFFFFFFFFFFFFF_4242FFFFFFFFFFFF_42FFFF42_V0102_SSERIAL_PPROPRIETARY"),
+				Expected: LoRaAllianceTR005Draft3{
+					JoinEUI:              types.EUI64{0x42, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff},
+					DevEUI:               types.EUI64{0x42, 0x42, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff},
+					VendorID:             [2]byte{0x42, 0xff},
+					ModelID:              [2]byte{0xff, 0x42},
+					DeviceValidationCode: "0102",
+					SerialNumber:         "SERIAL",
+					Proprietary:          "PROPRIETARY",
 				},
 			},
 			{
 				Name:          "EmptyExtensions",
-				Data:          []byte("LW:D0:42FFFFFFFFFFFFFF:4242FFFFFFFFFFFF:42FFFF42:O:S:P"),
-				CanonicalData: []byte("LW:D0:42FFFFFFFFFFFFFF:4242FFFFFFFFFFFF:42FFFF42"),
-				Expected: LoRaAllianceTR005{
-					JoinEUI:    types.EUI64{0x42, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff},
-					DevEUI:     types.EUI64{0x42, 0x42, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff},
-					VendorID:   [2]byte{0x42, 0xff},
-					ModelID:    [2]byte{0xff, 0x42},
-					OwnerToken: "",
+				Data:          []byte("URN:DEV:LW:42FFFFFFFFFFFFFF_4242FFFFFFFFFFFF_42FFFF42_V_S_P"),
+				CanonicalData: []byte("URN:DEV:LW:42FFFFFFFFFFFFFF_4242FFFFFFFFFFFF_42FFFF42"),
+				Expected: LoRaAllianceTR005Draft3{
+					JoinEUI:              types.EUI64{0x42, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff},
+					DevEUI:               types.EUI64{0x42, 0x42, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff},
+					VendorID:             [2]byte{0x42, 0xff},
+					ModelID:              [2]byte{0xff, 0x42},
+					DeviceValidationCode: "",
 				},
 			},
 			{
@@ -116,21 +115,28 @@ func TestLoRaAllianceTR005(t *testing.T) {
 			},
 			{
 				Name: "Invalid/Parts",
-				Data: []byte("LW:D0:42FFFFFFFFFFFFFF:4242FFFFFFFFFFFF"),
+				Data: []byte("URN:DEV:LW:42FFFFFFFFFFFFFF_4242FFFFFFFFFFFF"),
 				ErrorAssertion: func(t *testing.T, err error) bool {
 					return assertions.New(t).So(errors.IsInvalidArgument(err), should.BeTrue)
 				},
 			},
 			{
 				Name: "Invalid/EUI",
-				Data: []byte("LW:D0:42FFFFFFFF:4242FFFFFFFFFFFF:42FFFF42"),
+				Data: []byte("URN:DEV:LW:42FFFFFFFF_4242FFFFFFFFFFFF_42FFFF42"),
 				ErrorAssertion: func(t *testing.T, err error) bool {
 					return assertions.New(t).So(errors.IsInvalidArgument(err), should.BeTrue)
 				},
 			},
 			{
 				Name: "Invalid/ProdID",
-				Data: []byte("LW:D0:42FFFFFFFFFFFFFF:4242FFFFFFFFFFFF:42FFFF42AABB"),
+				Data: []byte("URN:DEV:LW:42FFFFFFFFFFFFFF_4242FFFFFFFFFFFF_42FFFF42AABB"),
+				ErrorAssertion: func(t *testing.T, err error) bool {
+					return assertions.New(t).So(errors.IsInvalidArgument(err), should.BeTrue)
+				},
+			},
+			{
+				Name: "Invalid/ExtensionChars",
+				Data: []byte("URN:DEV:LW:42FFFFFFFFFFFFFF_4242FFFFFFFFFFFF_42FFFF42_P#"),
 				ErrorAssertion: func(t *testing.T, err error) bool {
 					return assertions.New(t).So(errors.IsInvalidArgument(err), should.BeTrue)
 				},
@@ -139,7 +145,7 @@ func TestLoRaAllianceTR005(t *testing.T) {
 			t.Run(tc.Name, func(t *testing.T) {
 				a := assertions.New(t)
 
-				var data LoRaAllianceTR005
+				var data LoRaAllianceTR005Draft3
 				err := data.UnmarshalText(tc.Data)
 				if tc.ErrorAssertion != nil {
 					a.So(tc.ErrorAssertion(t, err), should.BeTrue)
