+++ pkg/qrcodegenerator/qrcodegenerator.go
@@ -22,6 +22,8 @@ import (
 	"go.thethings.network/lorawan-stack/v3/pkg/component"
 	"go.thethings.network/lorawan-stack/v3/pkg/errors"
 	"go.thethings.network/lorawan-stack/v3/pkg/log"
+	"go.thethings.network/lorawan-stack/v3/pkg/qrcodegenerator/qrcode"
+	"go.thethings.network/lorawan-stack/v3/pkg/qrcodegenerator/qrcode/enddevice"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 	"google.golang.org/grpc"
 )
@@ -33,6 +35,8 @@ type QRCodeGenerator struct {
 	*component.Component
 	ctx context.Context
 
+	qrCode *qrcode.QRCode
+
 	grpc struct {
 		endDeviceQRCodeGenerator *endDeviceQRCodeGeneratorServer
 		qrCodeParser             *qrCodeParserServer
@@ -43,12 +47,16 @@ var errFormatNotFound = errors.DefineNotFound("format_not_found", "format `{id}`
 
 // New returns a new *QRCodeGenerator.
 func New(c *component.Component, conf *Config) (*QRCodeGenerator, error) {
+	ctx := log.NewContextWithField(c.Context(), "namespace", "qrcodegenerator")
 	qrg := &QRCodeGenerator{
 		Component: c,
-		ctx:       log.NewContextWithField(c.Context(), "namespace", "qrcodegenerator"),
+		ctx:       ctx,
 	}
 	qrg.grpc.endDeviceQRCodeGenerator = &endDeviceQRCodeGeneratorServer{QRG: qrg}
 	qrg.grpc.qrCodeParser = &qrCodeParserServer{QRG: qrg}
+
+	qrCode := qrcode.New(ctx)
+	qrCode.RegisterEndDeviceFormat("tr005", new(enddevice.LoRaAllianceTR005Format))
 
 	c.RegisterGRPC(qrg)
 	return qrg, nil
